requirements.txt
celery==5.2.6
django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
gunicorn==20.1.0
django-heroku==0.3.1